+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Cleanup dashboard.
 * Add inline script.
 * Remove theme support.
+* Allow to undergister patterns.
 
 ## 0.5.4 (2023-06-26)